Moving tests from tests/test_repositories.py to tests/test_repository_pypi.py
